PDM_Vibration_Data/requirements.txt
fastapi==0.62.0
python-decouple==3.3
redis
fastapi-jwt-auth
uvicorn
keras
tensorflow
librosa
aiofiles
gunicorn
bz2file
pydub
matplotlib
opencv-python
numpy
pandas
seaborn
tqdm